comments and fix build error

+++ barretenberg/cpp/src/barretenberg/ultra_honk/decider_keys.hpp
@@ -117,7 +117,8 @@ template <typename Flavor_, size_t NUM_ = 2> struct DeciderVerificationKeys_ {
     {
         size_t max_log_circuit_size{ 0 };
         for (auto key : _data) {
-            max_log_circuit_size = std::max(max_log_circuit_size, key->verification_key->log_circuit_size);
+            max_log_circuit_size =
+                std::max(max_log_circuit_size, static_cast<size_t>(key->verification_key->log_circuit_size));
         }
         return max_log_circuit_size;
     }
